requirements.txt
DrissionPage~=4.0.5.6
requests~=2.32.3
beautifulsoup4~=4.12.3
loguru~=0.7.2
pandas~=2.2.2
typing~=3.7.4.3
pymongo~=4.12.0
PyYAML~=6.0.1
pydantic~=2.10.4
apscheduler
